refactor: move get-component-constructor to engine-dom
